bitcampTODO/requirements.txt
absl-py==1.4.0
anyio==3.7.0
asgiref==3.6.0
astunparse==1.6.3
cachetools==5.3.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Django==4.2.1
django-cors-headers==4.1.0
django-filter==23.2
djangorestframework==3.14.0
fastapi==0.98.0
flatbuffers==23.5.26
gast==0.4.0
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.0
h11==0.14.0
h5py==3.9.0
httptools==0.5.0
idna==3.4
keras==2.13.1rc0
libclang==16.0.0
Markdown==3.4.3
markdown2==2.4.8
MarkupSafe==2.1.3
numpy==1.25.0
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.9
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
sqlparse==0.4.4
starlette==0.27.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0rc1
tensorflow-estimator==2.13.0rc0
tensorflow-macos==2.13.0rc1
termcolor==2.3.0
typing_extensions==4.6.3
urllib3==1.26.16
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.3.6
wrapt==1.15.0
